enhance java client
