+++ CMakeLists.txt
@@ -502,10 +502,6 @@ macro(nebula_link_libraries target)
 endmacro(nebula_link_libraries)
 
 if (NOT SKIP_JAVA_CLIENT)
-    if (NOT EXISTS "${NEBULA_THIRDPARTY_ROOT}/fbthrift/thrift-1.0-SNAPSHOT.jar")
-        message(FATAL_ERROR "Doesn't have thrift-1.0-SNAPSHOT.jar")
-    endif()
-    set(java-fbthrift-jar ${NEBULA_THIRDPARTY_ROOT}/fbthrift/thrift-1.0-SNAPSHOT.jar)
     set(NEBULA_CLEAN_ALL_DEPS clean-interface clean-pch clean-hbase clean-java-client)
 else()
     set(NEBULA_CLEAN_ALL_DEPS clean-interface clean-pch clean-hbase)
