+++ src/client/java/README.md
@@ -1,0 +1,8 @@
+
+### Java Client
+
+
+The Java Client is dependent on `com.facebook:thrift:1.0.0-SNAPSHORT`.
+
+Please make sure the dependency have been install on your maven repo.
+
