+++ src/client/java/build.sh
@@ -1,14 +1,17 @@
 #!/bin/bash
- 
+
+# Copyright (c) 2018 vesoft inc. All rights reserved.
+#
+# This source code is licensed under Apache 2.0 License,
+# attached with Common Clause Condition 1.0, found in the LICENSES directory.
+
 #-----------------------------------------------
 # variable value
 #-----------------------------------------------
-# $1: the path of thrift-1.0-SNAPSHOT.jar
-# $2: the path of java src
-# $3: the path of current build dir
-java_fbthrift_jar=$1
-src_root_path=$2
-output_src_file=$3
+# $1: the path of java src
+# $2: the path of current build dir
+src_root_path=$1
+output_src_file=$2
 pom_file_path=${src_root_path}/pom.xml
 graph_gen_java_path=${output_src_file}/../../interface/gen-java/com/vesoft/nebula/graph/
 java_client_source_file_path=${src_root_path}/src/main/java/com/vesoft/nebula/graph/
@@ -30,7 +33,6 @@ check_file_exist()
 #-----------------------------------------------
 # check maven tool exist
 #-----------------------------------------------
-
 maven_check()
 {
 	command -v mvn >/dev/null 2>&1 || { echo >&2 "Require maven but it's not installed.  Aborting."; exit 1; }
@@ -51,9 +53,8 @@ setup_graph_source()
 #-----------------------------------------------
 compile_java_client()
 {
-	mvn clean package -DJAVA_FBTHRIFT_JAR=$java_fbthrift_jar -f ${pom_file_path}
+	mvn clean package -f ${pom_file_path}
 }
-
 
 setup_graph_source
 compile_java_client
