+++ src/client/java/src/main/java/com/vesoft/nebula/graph/client/ConnectionException.java
@@ -1,0 +1,19 @@
+/* Copyright (c) 2019 vesoft inc. All rights reserved.
+ *
+ * This source code is licensed under Apache 2.0 License,
+ * attached with Common Clause Condition 1.0, found in the LICENSES directory.
+ */
+
+package com.vesoft.nebula.graph.client;
+
+public class ConnectionException extends Exception {
+
+    public ConnectionException() {
+        super();
+    }
+
+    public ConnectionException(String message) {
+        super(message);
+    }
+}
+
