+++ src/tools/importer/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_custom_target(
     importer ALL
-    COMMAND "${CMAKE_CURRENT_SOURCE_DIR}/build.sh" ${java-fbthrift-jar} ${CMAKE_CURRENT_SOURCE_DIR} ">" "./_build.log" "2>&1"
+    COMMAND "${CMAKE_CURRENT_SOURCE_DIR}/build.sh" ${CMAKE_CURRENT_SOURCE_DIR} ">" "./_build.log" "2>&1"
     DEPENDS java-client
 )
